Crate dams

Cryptography, protocols, and other shared types and context used by multiple entities in the Lock Keeper digital asset management system.

⚠️ Usage: This crate should not be used directly by developers outside of Bolt Labs. It is designed as a dependency for other crates in the ecosystem, including dams-key-server and dams-client.

Re-exports

pub use error::DamsError;

Modules

audit_log

Audit log entries, fields, and types

blockchain

Blockchain abstraction.

channel

config

crypto

Application-specific cryptographic types and operations.

dams_rpc

defaults

error

keys

Digital asset keys and descriptions.

opaque_storage

pem_utils

timeout

transaction

Library for blockchains, transactions, and requests for signatures on transactions.

types

user

Models for the first pass of MongoDB integration.

Macros

impl_message_conversion

Generates TryFrom implementations to and from the Message type for a given list of types.

Enums

ClientAction

Options for actions the DAMS client can take.

RetrieveContext

Options for the asset owner’s intended use of a secret

TestLogs

Logs used to verify that an operation completed in the integration tests.