Module dams::keys

Digital asset keys and descriptions.

Includes basic key types (both standard keys and shares of keys) and identifiers and modifiers describing access control and custody.

Structs

Delegated

Use permission that delegates signing authority to a specified delegated party. The asset owner can set a UserPolicySpecification to apply additional rules and requirements for digital asset key usage.

DigitalAssetPublicKey

Public key portion of a digital asset key pair.

KeyInfo

Convenient grouping of the non-secret components of a digital asset key pair.

KeyMaterial

Key material that is generated externally to the system.

Passive

Use permission that refers to a digital asset key that was created by the Service Provider on behalf of an asset owner, i.e., the asset owner does not have a registered account with the key server. The Service Provider has custody of the key, including complete signing authority, i.e., there is no UserPolicySpecification set.

SelfCustodial

Use permission that assigns key custody to the asset owner. This permission requires an asset owner to actively participate in any action involving the digital asset key by authenticating with their password. The asset owner can set a UserPolicySpecification to apply additional rules and requirements for digital asset key usage.

SharedControl

Use restriction that assigns veto power to a given set of asset fiduciaries.

Unilateral

Use restriction that does not assign veto power to a given set of asset fiduciaries; that is, given a valid, authenticated request to use a digital asset key, no additional parties are consulted.

UserPolicySpecification

Specification of a user-set policy. This describes the set of policies that an asset owner can define around use of a digital asset key that are enforced by the DAMS key server. When a user-set policy is triggered, the key server can seek additional authentication or confirmation directly from the asset owner.

Traits

UsePermission

A use permission is a type that defines the degree to which the asset owner has delegated custody of a digital asset key or key share to another entity.

UseRestriction

A use restriction is a type that defines what entities have veto power over the use of a digital asset key.